2.0 Claude/requirements.txt
# Bird Motion Video Processor — Dependencies
#
# IMPORTANT: Use opencv-contrib-python for OpenCL GPU acceleration.
# Do NOT install opencv-python alongside it.
#
# Install with:
#   pip install -r requirements.txt

# OpenCV with OpenCL support (works on NVIDIA, AMD, Intel via standard drivers)
# Uninstall any other opencv first: pip uninstall opencv-python opencv-python-headless -y
opencv-contrib-python>=4.8.0

# Core
numpy>=1.24.0
PyQt5>=5.15.0
requests>=2.31.0

# Real-time folder monitoring
watchdog>=3.0.0

# YouTube upload (optional)
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0

# System monitoring
psutil>=5.9.0

# Optional AI bird detection
# torch>=2.0.0
# torchvision>=0.15.0
# ultralytics>=8.0.0